setup log in branch
