+++ PAL/App.js
@@ -4,12 +4,15 @@ import Dimensions from 'Dimensions';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import { Header, Button } from 'react-native-elements';
-import { createBottomTabNavigator } from 'react-navigation';
+import { createSwitchNavigator, createBottomTabNavigator } from 'react-navigation';
 import { AnimatedCircularProgress } from 'react-native-circular-progress';
 import CalendarStrip from 'react-native-calendar-strip';
 import moment from 'moment';
 import * as firebase from 'firebase';
-
+import Loading from './Loading'
+import SignUp from './SignUp'
+import Login from './Login'
+import Main from './main'
 import styles from './styles/home';
 import ScheduleScreen from './Schedule';
 import ReportsScreen from './Report';
@@ -17,296 +20,34 @@ import { GetGradient } from './gradient';
 import Modal from 'react-native-modal';
 import Slider from "react-native-slider";
 
-// Initialize Firebase
-const firebaseConfig = {
-    apiKey: "AIzaSyDA_RXtRHQI4IlCK-M2r9wgyBMYBFgs4m4",
-    authDomain: "pal394-a6f1f.firebaseapp.com",
-    databaseURL: "https://pal394-a6f1f.firebaseio.com",
-    projectId: "pal394-a6f1f",
-    storageBucket: "",
-    messagingSenderId: "33475295035"
-};
+//Initialize Firebase
+// const firebaseConfig = {
+//     apiKey: "AIzaSyDA_RXtRHQI4IlCK-M2r9wgyBMYBFgs4m4",
+//     authDomain: "pal394-a6f1f.firebaseapp.com",
+//     databaseURL: "https://pal394-a6f1f.firebaseio.com",
+//     projectId: "pal394-a6f1f",
+//     storageBucket: "",
+//     messagingSenderId: "33475295035"
+// };
 
-var curfirebase=firebase.initializeApp(firebaseConfig);
-var addressfirebase=curfirebase.database().ref("/Jason/address");
-var statusfirebase=curfirebase.database().ref("/Jason/status");
-var batfirebase=curfirebase.database().ref("/Jason/battery");
-
-class HomeScreen extends React.Component {
-  constructor(props){
-    super(props)
-    this.state = {
-    progress: 0.5,
-    statusColor: "#6fd865",
-    statusEmoji: require("./resources/smile.png"),
-    opacity: 0.5,
-    battery: 0.0,
-    lat: "",
-    lng: "",
-    address: ""
-    }
-  }
-
-  componentDidMount() {
-    addressfirebase.on('value', snapshot => {this.setState({address: snapshot.val()})
-    });
-    statusfirebase.on('value', snapshot => {
-    	this.setState({progress: snapshot.val()});
-    	this._updateStatus(snapshot.val());
-    });
-    batfirebase.on('value', snapshot => {this.setState({battery: snapshot.val()})
-    });
-  }
-
-  _updateStatus = status =>{
-
-  	//change emoji
-  	var img;
-
-  	if(status < 0.2){
-  		var img = require('./resources/sad.png');
-  	}else if(status < 0.4){
-  		img = require('./resources/angry.png');
-  	}else if(status < 0.6){
-  		img = require('./resources/normal.png');
-  	}else if(status < 0.8){
-  		var img = require('./resources/smile.png');
-  	}else{
-  		var img = require('./resources/happy.png');
-  	}
-		this.setState({statusEmoji: img});
-  }
-
-  _getGradient(ratio){
-  			ratio = 1 - ratio;
-	  // var red =     '#ff0000';
-	  // var yellow =  '#ffff00';
-	  // var green =   '#008000';
-	  if (ratio == 0.5){
-	    return "#ffff00";
-	  }
-
-	  else if(ratio < 0.5){
-	    let r = ratio/0.5;
-	    let r2 = 1-r;
-
-	    let red = Math.round(255*r).toString(16); // #ff
-	    let green = Math.round(255*r + 128*r2).toString(16); // #ff*r + #80*r2
-	    // blue is 00
-	    if (red.length == 1) red = '0'+red;
-	    if (green.length == 1) green = '0' + green;
-
-	    return '#'+red+green+"00";
-	  }
-	  else {
-	    let r2 = 1-((ratio-.5)/.5);
-
-	    let green = Math.round(255*r2).toString(16);
-	    if (green.length == 1) green = '0' + green;
-	    return '#ff'+green+"00";
-	  }
-	}
-
-  state = {
-    visibleModal: null,
-    sliderValue: 0.2,
-  };
-
-  _renderButton = (text, onPress) => (
-    <TouchableOpacity onPress={onPress}>
-      <View style={styles.meltdownButton}>
-        <Text>{text}</Text>
-      </View>
-    </TouchableOpacity>
-  );
-
-  _renderModalContent = () => (
-    <View style={styles.modalContent}>
-      <Text style={{fontWeight: 'bold'}}>Record meltdown</Text>
-      <Text>Current time: {new Date().toLocaleString()}</Text>
-      <View style={styles.sliderContainer}>
-        <Slider
-          value={this.state.sliderValue}
-          onValueChange={value => this.setState({ value })}
-        />
-        <Text>Value: {this.state.sliderValue}</Text>
-      </View>
-      {this._renderButton('Close', () => this.setState({ visibleModal: null }))}
-    </View>
-  );
-
-   render() {
-    return (
-        <View style={{flex:1}}>
-          <View>
-            <Header
-            placement="left"
-            backgroundColor = "#ff1900"
-            leftComponent={< ShirtStatus />}
-            centerComponent={{ text: 'PAL', style: {color: '#fff', marginLeft: -30} }}
-            rightComponent={{ icon: 'menu', color: '#fff' }}
-            />
-          <View>
-            <Text style={styles.statusTitle}>{"Jason's Current Status"}</Text>
-            <Text style={styles.statusGreen}>Green</Text>
-            <AnimatedCircularProgress
-              style = {styles.semiCircleContainer}
-              size={Dimensions.get('window').width-100}
-              width={25}
-              fill={100*this.state.progress}
-              arcSweepAngle={180}
-              rotation={270}
-              tintColor= {GetGradient(this.state.progress)}
-              backgroundColor="#666"
-              onAnimationComplete={() => console.log('onAnimationComplete')}
-            />
-            <View style={{ flex:1, justifyContent: 'center', alignItems: 'center'}}>
-              <Image style={styles.face}
-                source={this.state.statusEmoji}
-              />
-            </View>
-              <View style={{ justifyContent: 'center', top: 175, height: 75, backgroundColor: '#e4e4e4'}}>
-              <Text style = {{textAlign: 'center'}}> Jason is at {this.state.address}</Text>
-              </View>
-          </View>
-          </View>
-          <FlatList style={styles.flatListContainer}
-              data={data}
-              renderItem={({item}) => (
-                <TouchableHighlight onPress={() => {}}
-                  activeOpacity={this.state.opacity}
-                  underlayColor="#fff">
-                  <View style={styles.itemContainer}>
-                    <Icon style={styles.searchIcon} name={item.icon} size={20} color="#000" />
-                    <Text style={styles.item}>{item.name}</Text>
-                  </View>
-                </TouchableHighlight>
-              )}
-              keyExtractor={item => item.id}
-              numColumns={numColumns}
-            />
-            <View style={styles.meltdownContainer}>
-              {this._renderButton('Record Meltdown', () => this.setState({ visibleModal: 1 }))}
-              <Modal isVisible={this.state.visibleModal === 1}>
-                {this._renderModalContent()}
-              </Modal>
-            </View>
-        </View>
-    );
-  }
-}
-
-
-class ShirtStatus extends React.Component{
-  constructor(props){
-    super(props)
-    this.state = {
-
-    battery: 0.0,
-
-    }
-  }
-  componentDidMount() {
-
-    batfirebase.on('value', snapshot => {this.setState({battery: snapshot.val()})
-    });
-  }
-  renderBattery() {
-    if (this.state.battery>.875) {
-      return <Text>
-        <Image
-          source={require('./resources/tshirt_white.png')}
-          style= {{height: 20, width: 20}}
-        />
-        <Icon name={"battery-full"} size={20} color="#fff"/>
-      </Text>
-    }
-    else if (this.state.battery>.625) {
-      return <Text>
-        <Image
-          source={require('./resources/tshirt_white.png')}
-          style= {{height: 20, width: 20}}
-        />
-        <Icon name={"battery-three-quarters"} size={20} color="#fff"/>
-      </Text>
-    }
-    else if (this.state.battery>.375) {
-      return <Text>
-        <Image
-          source={require('./resources/tshirt_white.png')}
-          style= {{height: 20, width: 20}}
-        />
-        <Icon name={"battery-half"} size={20} color="#fff"/>
-      </Text>
-    }
-     else if (this.state.battery>.125) {
-      return <Text>
-        <Image
-          source={require('./resources/tshirt_white.png')}
-          style= {{height: 20, width: 20}}
-        />
-        <Icon name={"battery-quarter"} size={20} color="#fff"/>
-      </Text>
-    }
-      else {
-        return <Text>
-        <Image
-          source={require('./resources/tshirt_white.png')}
-          style= {{height: 20, width: 20}}
-        />
-        <Icon name={"battery-empty"} size={20} color="#fff"/>
-      </Text>
-      }
+// var curfirebase=firebase.initializeApp(firebaseConfig);
+// var addressfirebase=curfirebase.database().ref("/Jason/address");
+// var statusfirebase=curfirebase.database().ref("/Jason/status");
+// var batfirebase=curfirebase.database().ref("/Jason/battery");
 
 
 
-  }
-  render() {
-    return (
-            <View >
-     {this.renderBattery()}
-        </View>
-
-    );
-  }
-};
-
-
-export default createBottomTabNavigator(
+const App = createSwitchNavigator(
   {
-    Home: HomeScreen,
-    Report: ReportsScreen,
-    Schedule: ScheduleScreen,
+    Loading,
+    SignUp,
+    Login,
+    Main
   },
   {
-    navigationOptions: ({ navigation }) => ({
-      tabBarIcon: ({ focused, tintColor }) => {
-        const { routeName } = navigation.state;
-        let iconName;
-        if (routeName === 'Home') {
-          iconName = `ios-information-circle${focused ? '' : '-outline'}`;
-        } else if (routeName === 'Report') {
-          iconName = `ios-options${focused ? '' : '-outline'}`;
-        }
-        else if (routeName === 'Schedule') {
-          iconName = `ios-people${focused ? '' : '-outline'}`;
-        }
-        return <Ionicons name={iconName} size={25} color={tintColor} />;
-      },
-    }),
-    tabBarOptions: {
-      activeTintColor: 'tomato',
-      inactiveTintColor: 'gray',
-    },
+    initialRouteName: 'Loading'
   }
 )
 
-const data = [
-  {id: 1, name: 'Message', icon: 'comments'},
-  {id: 2, name: 'History', icon: 'bar-chart'},
-  {id: 3, name: 'Share', icon:'share'},
-  {id: 4, name: 'Heart Rate', icon:'heart'},
-  {id: 5, name: 'Send a Hug or Calming Technique'},
-  {id: 6, name: 'Body Temperature', icon: 'thermometer-0'},
-];
-const numColumns = 3;
+export default App
+
