+++ PAL/Loading.js
@@ -1,0 +1,27 @@
+// Loading.js
+import React from 'react'
+import * as firebase from 'firebase';
+
+import { View, Text, ActivityIndicator, StyleSheet } from 'react-native'
+export default class Loading extends React.Component {
+    componentDidMount() {
+    firebase.auth().onAuthStateChanged(user => {
+      this.props.navigation.navigate(user ? 'Main' : 'SignUp')
+    })
+  }
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text>Loading</Text>
+        <ActivityIndicator size="large" />
+      </View>
+    )
+  }
+}
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  }
+})
